Move intervals track to an extension to reduce main bundle size
Intervals track is used by a small subset of LZ installations,
 primarily the T2D Portal.
