+++ examples/interval_annotations.html
@@ -7,9 +7,12 @@
     <link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/skeleton/2.0.4/skeleton.css"/>
 
     <!-- Necessary includes for LocusZoom.js -->
+    <link rel="stylesheet" href="../dist/locuszoom.css" type="text/css"/>
     <script src="../dist/locuszoom.vendor.min.js" type="text/javascript"></script>
     <script src="../dist/locuszoom.app.min.js" type="text/javascript"></script>
-    <link rel="stylesheet" href="../dist/locuszoom.css" type="text/css"/>
+
+      <!-- Helpers that use the external library to add functionality to LocusZoom -->
+  <script src="../dist/ext/lz-intervals-track.min.js" type="application/javascript"></script>
 
     <title>LocusZoom.js ~ Interval Annotations Example</title>
